feat: add hasBodyPayload method to ClientRequest and HTTPClientRequest interface
