+++ src/Contracts/HTTPClientRequest.php
@@ -57,6 +57,7 @@ interface HTTPClientRequest
     public function hasQuery(): bool;
     public function hasSecurityContext(): bool;
     public function hasBasicAuth(): bool;
+    public function hasBodyPayload(): bool;
 
     public function setMethod(string $method): self;
     public function setUri(string $uri): self;
